backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
requests==2.32.3
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
argon2-cffi==23.2.0